Refactor to make Batch Update more like Batch Upsert
